修改某些引用错误和播放器状态错误

+++ app/src/main/java/com/zihany/cloudmusic/song/bean/SongDetailBean.kt
@@ -45,7 +45,7 @@ data class Song(
     val id: Int,
     val l: L,
     val m: M,
-    val mark: Int,
+    val mark: Long,
     val mst: Int,
     val mv: Int,
     val name: String,
